Refactor: saparate cron jobs
